make editor constants global, move Output files out of Editor, moved button components to common, remove unnecessary button imports
