--- src/components/Editor/constants/index.js
@@ -1,7 +1,0 @@
-//View Mode
-export const CODE_AND_OUTPUT = 0;
-export const CODE_ONLY = 1;
-export const OUTPUT_ONLY = 2;
-
-//UI
-export const EDITOR_WIDTH_BREAKPOINT = 1000;
