+++ src/components/Sketches/index.js
@@ -3,9 +3,8 @@ import SketchBox from "./components/SketchBox";
 import ConfirmDeleteModalContainer from "./containers/ConfirmDeleteModalContainer";
 import CreateSketchModalContainer from "./containers/CreateSketchModalContainer";
 import EditSketchModalContainer from "./containers/EditSketchModalContainer";
-import OpenPanelButtonContainer from "../common/containers/OpenPanelButtonContainer";
 import { SketchThumbnailArray } from "./constants";
-import ProcessingConstructor from "../Editor/components/Output/Processing";
+import ProcessingConstructor from "../Output/Processing";
 // import { PANEL_SIZE } from "../../constants";
 import "styles/Sketches.scss";
 
@@ -89,7 +88,6 @@ class Sketches extends React.Component {
 
   renderHeader = () => (
     <div className="sketches-header">
-      <OpenPanelButtonContainer />
       <div className="sketches-header-text">Sketches</div>
       <Button
         className="ml-auto mr-2"
