+++ src/constants/index.js
@@ -66,6 +66,15 @@ if (process && process.env) {
 
 const PANEL_SIZE = 250;
 
+// Editor and Output constants
+//View Mode
+const CODE_AND_OUTPUT = 0;
+const CODE_ONLY = 1;
+const OUTPUT_ONLY = 2;
+
+//UI
+const EDITOR_WIDTH_BREAKPOINT = 1000;
+
 module.exports = {
   //Language definitions
   PYTHON,
@@ -107,6 +116,14 @@ module.exports = {
 
   //codemirror conversions
   CODEMIRROR_CONVERSIONS,
+
+  // editor constants:
+  CODE_AND_OUTPUT,
+  CODE_ONLY,
+  OUTPUT_ONLY,
+
+  // UI
+  EDITOR_WIDTH_BREAKPOINT,
 
   //Firebase constants
   EMAIL_DOMAIN_NAME: "@fake.com",
